Changed Graph class so that Graph object contains only information relevant for the genealogy graph it represents.
Previously, a Graph object contained edge information for all edges
associated with a node in the Math Genealogy Project's information.
This included edges to nodes that are not part of the genealogy
graph constructed. This change ensures that only edge information
for edges between nodes in the genealogy graph is retained.

This closes #2.