Generalize grabbing interface.
This change modifies the Geneagrapher class so that it allows a caller
to provide a custom grabber class. The advantage of this is that it
allows the introduction of more complex grabbers, such as grabbers that
provide caching.

To accomplish this, the Grabber class is being modified to work across
grabbing session (previously a Grabber object was used for retrieving
the record of a single mathematician), which is necessary for a
cache-based grabber. Additionally, the Grabber class is now a context
manager, which allows using it with the Python "with" statement.
Although, this provides no real advantage for Grabber, it will be
advantageous for a grabber that opens a resource when initialized and
closes the resource when it falls out of scope.

This change closes #7.