Add LocalDataGrabber class so tests work with local data.
This changeset adds the LocalDataGrabber class in the test source
tree. This class exposes the same interface as Grabber -- so that
it can be used in place of it -- but is only able to grab test
data from the local test data directory. This is useful for testing
without making network requests.