Adding CacheGrabber -- a grabber class that caches results.
The CacheGrabber uses the standard library's shelve module to persist
cached records. The cache stores all of the information extracted
from a Math Genealogy Project web page, in addition to a timestamp
that records when the record was obtained. The timestamp is used to
determine when a cached value is invalid due to being stale.

This changeset is part of #8.