Update test to use large cache expiration timeout.
A test meant to verify the behavior of the Geneagrapher when using
CacheGrabber was failing because the cache entry had gone stale (it
happens after one week by default). To keep the test from starting
to fail after one week, this commit modifies the test to use a
very long cache item expiration interval, effectively ensuring the
cache item will not go stale.