Replace relative image paths with URL to tagged image
This is being done so that images will work in the 2.0.0 release
published to PyPI. Images do not load from a README in PyPI if they
are relative paths to files in the repository, so a URL is being
used instead. Moreover, the URL contains the v2.0.0 tag in the
path. This will ensure that even if the images change or are
deleted in the future, the published version will display the
images from publication time.

These changes are only intended for PyPI, so the commit will be
rolled back beyond the tag commit.